+++ src/main/java/com/jpdev/repository/UserRepository.java
@@ -2,10 +2,9 @@ package com.jpdev.repository;
 
 import com.jpdev.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
 }
